First success

+++ cfast_slic.pxd
@@ -1,6 +1,6 @@
 # cython: language_level=3
 
-from libc.stdint cimport uint8_t, uint32_t, uint16_t, int32_t
+from libc.stdint cimport uint8_t, uint32_t, uint16_t, int16_t
 
 cdef extern from "fast-slic-common.h":
     ctypedef struct Cluster:
@@ -22,16 +22,16 @@ cdef extern from "fast-slic-common.h":
 
 cdef extern from "fast-slic.h":
     void fast_slic_initialize_clusters(int H, int W, int K, const uint8_t* image, Cluster *clusters) nogil
-    void fast_slic_iterate(int H, int W, int K, float compactness, float min_size_factor, uint8_t quantize_level, int max_iter, const uint8_t* image, Cluster* clusters, uint32_t* assignment) nogil
-    Connectivity* fast_slic_get_connectivity(int H, int W, int K, const uint32_t *assignment) nogil
+    void fast_slic_iterate(int H, int W, int K, float compactness, float min_size_factor, uint8_t quantize_level, int max_iter, const uint8_t* image, Cluster* clusters, uint16_t* assignment) nogil
+    Connectivity* fast_slic_get_connectivity(int H, int W, int K, const uint16_t *assignment) nogil
     Connectivity* fast_slic_knn_connectivity(int H, int W, int K, const Cluster* clusters, int num_neighbors) nogil
     void fast_slic_free_connectivity(Connectivity* conn) nogil
-    void fast_slic_get_mask_density(int H, int W, int K, const Cluster* clusters, const uint32_t* assignment, const uint8_t *mask, uint8_t *cluster_densities) nogil
-    void fast_slic_cluster_density_to_mask(int H, int W, int K, const Cluster *clusters, const uint32_t* assignment, const uint8_t *cluster_densities, uint8_t *result) nogil
+    void fast_slic_get_mask_density(int H, int W, int K, const Cluster* clusters, const uint16_t* assignment, const uint8_t *mask, uint8_t *cluster_densities) nogil
+    void fast_slic_cluster_density_to_mask(int H, int W, int K, const Cluster *clusters, const uint16_t* assignment, const uint8_t *cluster_densities, uint8_t *result) nogil
 
 cdef extern from "fast-slic-avx2.h":
     void fast_slic_initialize_clusters_avx2(int H, int W, int K, const uint8_t* image, Cluster *clusters) nogil
-    void fast_slic_iterate_avx2(int H, int W, int K, float compactness, float min_size_factor, uint8_t quantize_level, int max_iter, const uint8_t* image, Cluster* clusters, uint32_t* assignment) nogil
+    void fast_slic_iterate_avx2(int H, int W, int K, float compactness, float min_size_factor, uint8_t quantize_level, int max_iter, const uint8_t* image, Cluster* clusters, uint16_t* assignment) nogil
     int fast_slic_supports_avx2() nogil
 
 
@@ -50,10 +50,10 @@ cdef class BaseSlicModel:
 
     cpdef void initialize(self, const uint8_t [:, :, ::1] image)
     cpdef iterate(self, const uint8_t [:, :, ::1] image, int max_iter, float compactness, float min_size_factor, uint8_t quantize_level)
-    cpdef get_connectivity(self, const int32_t[:,::1] assignments)
-    cpdef get_knn_connectivity(self, const int32_t[:,::1] assignments, size_t num_neighbors)
-    cpdef get_mask_density(self, const uint8_t[:, ::1] mask, const int32_t[:, ::1] assignments)
-    cpdef broadcast_density_to_mask(self, const uint8_t[::1] densities, const int32_t[:, ::1] assignments);
+    cpdef get_connectivity(self, const int16_t[:,::1] assignments)
+    cpdef get_knn_connectivity(self, const int16_t[:,::1] assignments, size_t num_neighbors)
+    cpdef get_mask_density(self, const uint8_t[:, ::1] mask, const int16_t[:, ::1] assignments)
+    cpdef broadcast_density_to_mask(self, const uint8_t[::1] densities, const int16_t[:, ::1] assignments);
     cdef _get_clusters(self)
     cdef _set_clusters(self, clusters)
